convert to ts, add test cases for reject and accept api calls.
